+++ Infrastructure/HotfixMods.Infrastructure/Services/GameobjectService.Miscellaneous.cs
@@ -11,8 +11,8 @@ namespace HotfixMods.Infrastructure.Services
         {
             // Step 1: Init IDs of single entities
             var hotfixModsEntityId = await GetIdByConditionsAsync<HotfixModsEntity>(dto.HotfixModsEntity.ID, dto.IsUpdate);
-            var gameobjectTemplateId = await GetIdByConditionsAsync<GameobjectTemplate>((int)dto.GameobjectTemplate.Entry, dto.IsUpdate);
-            var gameobjectDisplayInfoId = await GetIdByConditionsAsync<GameobjectDisplayInfo>(dto.GameobjectDisplayInfo.ID, dto.IsUpdate);
+            var gameobjectTemplateId = await GetIdByConditionsAsync<GameobjectTemplate>(dto.GameobjectTemplate.Entry, dto.IsUpdate);
+            var gameobjectDisplayInfoId = await GetIdByConditionsAsync<GameobjectDisplayInfo>((ulong)dto.GameobjectDisplayInfo.ID, dto.IsUpdate);
 
             // Step 2: Prepare IDs of list entities
             // Nothing to do here
@@ -26,7 +26,7 @@ namespace HotfixMods.Infrastructure.Services
             dto.GameobjectTemplate.VerifiedBuild = VerifiedBuild;
             dto.GameobjectTemplate.DisplayID = (uint)gameobjectDisplayInfoId;
 
-            dto.GameobjectDisplayInfo.ID = gameobjectDisplayInfoId;
+            dto.GameobjectDisplayInfo.ID = (int)gameobjectDisplayInfoId;
             dto.GameobjectDisplayInfo.VerifiedBuild = VerifiedBuild;
 
             if(dto.GameobjectTemplateAddon != null)
