+++ Infrastructure/HotfixMods.Infrastructure/Services/SpellVisualKitService.cs
@@ -58,7 +58,7 @@ namespace HotfixMods.Infrastructure.Services
                 var result = new SpellVisualKitDto()
                 {
                     SpellVisualKit = spellVisualKit,
-                    HotfixModsEntity = await GetExistingOrNewHotfixModsEntityAsync(spellVisualKit.ID),
+                    HotfixModsEntity = await GetExistingOrNewHotfixModsEntityAsync((ulong)spellVisualKit.ID),
                 };
 
                 var spellVisualKitEffects = await GetAsync<SpellVisualKitEffect>(callback, progress, new DbParameter(nameof(SpellVisualKitEffect.ParentSpellVisualKitID), id));
